refactor(#3038): VoiceBargeInRuntime S5 — extract live-cut playback cluster
S5 scope: extract the next cohesive VoiceBargeInRuntime subcluster from the S1
concern map without behavior changes. The S4 remaining-map left two cohesive
candidates: live-cut playback and the TTS pipeline. From the S1 map, live-cut
totals 108 playback lines (`reset_after_playback_start` 12,
`reset_after_playback_start_with_owner` 32, `clear_playback` 5,
`clear_playback_if_owner` 5, `observe_live_pcm_i16` 36, `monitor_for_channel`
13, `update_existing_monitor_sensitivity` 5) plus `current_sensitivity` (8) —
mapped lifecycle/config in S1 but reassigned by the S4 remaining-map as the
live-cut "sensitivity mirror": it is the monitor-sensitivity read path (F18
atomic mirror fallback) and two of its three callers move with the cluster.
The TTS pipeline totals 83 lines (`set_runtime_tts_voice` 24,
`synthesize_acknowledgement` 9, `synthesize_progress_tts` 50). Live-cut is the
larger candidate and moves cleanly, so it is the pick.

Cluster boundary: moved `reset_after_playback_start`,
`reset_after_playback_start_with_owner`, `clear_playback`,
`clear_playback_if_owner`, `observe_live_pcm_i16`, `monitor_for_channel`,
`current_sensitivity`, and `update_existing_monitor_sensitivity` verbatim into
`src/services/discord/voice_barge_in/live_cut_playback.rs` as an inherent impl
child module. The only moved-method signature edits are visibility-only:

- `reset_after_playback_start_with_owner`: private -> `pub(super)` (sibling
  callers in `progress_playback.rs` and `final_result_playback.rs`, plus root
  tests).
- `clear_playback_if_owner`: private -> `pub(super)` (same sibling callers,
  plus root tests).
- `current_sensitivity`: private -> `pub(super)` (root caller
  `handle_processing_transcript`).
- `update_existing_monitor_sensitivity`: private -> `pub(super)` (root callers
  `set_sensitivity` and `apply_voice_command`).
- `reset_after_playback_start`, `clear_playback`, `observe_live_pcm_i16` keep
  `pub(in crate::services::discord)`; `monitor_for_channel` stays private (its
  only callers move with the cluster).

No root `settings`-style import mirror was needed: the moved bodies contain no
`super::` paths and every referenced helper (`lock_monitor`,
`LivePlaybackSession`, `pcm_i16_to_le_bytes`, `LiveBargeInMonitor`,
`LiveBargeInCut`, `BargeInPlayerStop`, `BargeInSensitivity`) resolves through
the child's `use super::*`.

No await point, lock scope, spawn, map-entry, cancellation, channel-ordering,
or queue-mutation behavior changed (the whole cluster is sync; the
`lock_monitor` scopes and `playbacks`/`monitors` DashMap operations moved
byte-identical). No struct fields were widened. Existing cfg behavior is
unchanged.

LoC accounting: `voice_barge_in.rs` production LoC drops 3332 -> 3217 (-115).
New child `voice_barge_in/live_cut_playback.rs` is 120 production LoC, below
the giant threshold, so only the root ratchet is lowered.

Token proof: normalized method-token comparison against HEAD is identical
after visibility-only normalization. Per moved method:

| Method | Method tokens | Body tokens | Normalized SHA |
| --- | ---: | ---: | --- |
| `reset_after_playback_start` | 55 -> 55 | 15 -> 15 | `849291a2426f` |
| `reset_after_playback_start_with_owner` | 138 -> 138 | 98 -> 98 | `c72a3abf78d4` |
| `clear_playback` | 33 -> 33 | 16 -> 16 | `2da3ac0e2037` |
| `clear_playback_if_owner` | 45 -> 45 | 31 -> 31 | `805621e82981` |
| `observe_live_pcm_i16` | 222 -> 222 | 198 -> 198 | `153396d1caab` |
| `monitor_for_channel` | 81 -> 81 | 51 -> 51 | `91f5b8b858ab` |
| `current_sensitivity` | 18 -> 18 | 9 -> 9 | `75c5bf5fd79a` |
| `update_existing_monitor_sensitivity` | 35 -> 35 | 25 -> 25 | `21fa7891bfc6` |
| **Total** | **627 -> 627** | **443 -> 443** | |

Remaining-map status after S5:

- routing/dispatch still owns the heavier dispatch path in the root: channel
  text reply, processing transcript, transcript announcement, background
  handoff, start voice turn, completed utterance processing, deferred drain,
  and the inflight-foreground cancel registry.
- TTS pipeline remains in the root: runtime TTS voice rebuild plus
  progress/ack synthesis (`set_runtime_tts_voice`, `synthesize_acknowledgement`,
  `synthesize_progress_tts`).
- lifecycle/config and misc methods remain below the larger cohesive
  candidates.

Gates run before commit:

- cargo fmt --check
- cargo check --lib
- cargo clippy --lib -- -D warnings
- set -o pipefail; cargo test --lib voice 2>&1 | grep "test result" (277
  passed) && cargo test --lib discord 2>&1 | grep "test result" (1620 passed)
- python3 scripts/audit_maintainability.py --check
