+++ ARCHITECTURE.md
@@ -524,6 +524,7 @@ src/
 │   │   │   └── cleanup.rs
 │   │   ├── voice_barge_in/
 │   │   │   ├── final_result_playback.rs
+│   │   │   ├── live_cut_playback.rs
 │   │   │   ├── progress_playback.rs
 │   │   │   ├── routing.rs
 │   │   │   └── stt.rs
