+++ src/services/discord/voice_barge_in.rs
@@ -47,6 +47,8 @@ use super::{SharedData, http, mailbox_has_active_turn, rate_limit_wait, settings
 
 #[path = "voice_barge_in/final_result_playback.rs"]
 mod final_result_playback;
+#[path = "voice_barge_in/live_cut_playback.rs"]
+mod live_cut_playback;
 #[path = "voice_barge_in/progress_playback.rs"]
 mod progress_playback;
 #[path = "voice_barge_in/routing.rs"]
@@ -1768,96 +1770,6 @@ impl VoiceBargeInRuntime {
         Some(sensitivity)
     }
 
-    #[allow(dead_code)] // #3034: test-only playback-reset; prod uses reset_after_playback_start_with_owner
-    pub(in crate::services::discord) fn reset_after_playback_start<P>(
-        &self,
-        channel_id: ChannelId,
-        player: Arc<P>,
-        cancellation: CancellationToken,
-    ) where
-        P: BargeInPlayerStop + 'static,
-    {
-        self.reset_after_playback_start_with_owner(channel_id, player, cancellation, None);
-    }
-
-    fn reset_after_playback_start_with_owner<P>(
-        &self,
-        channel_id: ChannelId,
-        player: Arc<P>,
-        cancellation: CancellationToken,
-        owner: Option<u64>,
-    ) where
-        P: BargeInPlayerStop + 'static,
-    {
-        if !self.barge_in_enabled {
-            return;
-        }
-
-        let sensitivity = self.current_sensitivity();
-        let monitor = self.monitor_for_channel(channel_id, sensitivity);
-        {
-            let mut monitor = lock_monitor(&monitor);
-            monitor.set_sensitivity(sensitivity);
-            monitor.reset_after_playback_start();
-        }
-
-        let player: Arc<dyn BargeInPlayerStop> = player;
-        self.playbacks.insert(
-            channel_id.get(),
-            Arc::new(LivePlaybackSession {
-                player,
-                cancellation,
-                owner,
-            }),
-        );
-    }
-
-    #[allow(dead_code)] // #3034: test-only playback clear; prod uses clear_playback_if_owner
-    pub(in crate::services::discord) fn clear_playback(&self, channel_id: ChannelId) {
-        self.playbacks.remove(&channel_id.get());
-    }
-
-    fn clear_playback_if_owner(&self, channel_id: ChannelId, owner: u64) {
-        self.playbacks
-            .remove_if(&channel_id.get(), |_, session| session.owner == Some(owner));
-    }
-
-    pub(in crate::services::discord) fn observe_live_pcm_i16(
-        &self,
-        channel_id: ChannelId,
-        samples: &[i16],
-    ) -> Option<LiveBargeInCut> {
-        if !self.barge_in_enabled || samples.is_empty() {
-            return None;
-        }
-
-        let playback = self
-            .playbacks
-            .get(&channel_id.get())
-            .map(|entry| entry.value().clone())?;
-        let sensitivity = self.current_sensitivity();
-        let monitor = self.monitor_for_channel(channel_id, sensitivity);
-        let mut monitor = lock_monitor(&monitor);
-        monitor.set_sensitivity(sensitivity);
-
-        let pcm = pcm_i16_to_le_bytes(samples);
-        match monitor.observe_pcm(&pcm, playback.player.as_ref(), &playback.cancellation) {
-            Ok(Some(cut)) => {
-                self.playbacks.remove(&channel_id.get());
-                Some(cut)
-            }
-            Ok(None) => None,
-            Err(error) => {
-                tracing::warn!(
-                    error = %error,
-                    channel_id = channel_id.get(),
-                    "voice live barge-in stop failed"
-                );
-                None
-            }
-        }
-    }
-
     pub(in crate::services::discord) async fn handle_processing_transcript(
         &self,
         shared: &Arc<SharedData>,
@@ -3119,33 +3031,6 @@ impl VoiceBargeInRuntime {
             .entry(channel_id.get())
             .or_insert_with(|| Arc::new(Mutex::new(DeferredBargeInBuffer::new())))
             .clone()
-    }
-
-    fn monitor_for_channel(
-        &self,
-        channel_id: ChannelId,
-        sensitivity: BargeInSensitivity,
-    ) -> Arc<std::sync::Mutex<LiveBargeInMonitor>> {
-        self.monitors
-            .entry(channel_id.get())
-            .or_insert_with(|| {
-                Arc::new(std::sync::Mutex::new(LiveBargeInMonitor::new(sensitivity)))
-            })
-            .clone()
-    }
-
-    fn current_sensitivity(&self) -> BargeInSensitivity {
-        // F18 (#2046): try_read 실패 시 boot-time default 가 아닌 가장 최근에
-        // 설정된 sensitivity 를 반환하도록 atomic mirror 로 폴백한다. TTL reset
-        // 이 일어나는 짧은 윈도우라도 사용자가 설정한 Conservative 가 잠깐
-        // Normal 로 평가되는 회귀를 막는다.
-        self.sensitivity.current()
-    }
-
-    fn update_existing_monitor_sensitivity(&self, sensitivity: BargeInSensitivity) {
-        for monitor in &self.monitors {
-            lock_monitor(monitor.value()).set_sensitivity(sensitivity);
-        }
     }
 }
